refactor: Added backgrounds folder
