+++ src/backgrounds/background.h
@@ -1,8 +1,8 @@
 #ifndef BACKGROUND_H
 #define BACKGROUND_H
 
-#include "rt3.h"
-#include "paramset.h"
+#include "../core/rt3.h"
+#include "../core/paramset.h"
 
 namespace rt3 {
 // TODO: Create a virtual class Background and derive BackgroundColor,
